Centralizes MSVC POSIX compatibility definitions
Introduces a shared compatibility header to unify and simplify
POSIX-to-MSVC definitions and wrappers across the codebase,
replacing scattered local definitions. Improves maintainability
and reduces redundancy for Windows builds.
