+++ src/modules/gdk/producer_pixbuf.c
@@ -41,23 +41,7 @@
 #include <unistd.h>
 #endif
 #ifdef _MSC_VER
-    #include <io.h>      // For _close, _write
-    #include <BaseTsd.h> // For SSIZE_T
-
-    // Define ssize_t for MSVC
-    typedef SSIZE_T ssize_t;
-
-
-static inline int close(int fd)
-{
-    return _close(fd);
-}
-
-// write 的包装器
-static inline int write(int fd, const void *buffer, unsigned int count)
-{
-    return _write(fd, buffer, count);
-}
+#include <framework/msvc_posix_compat.h>
 #endif
 // this protects concurrent access to gdk_pixbuf
 static pthread_mutex_t g_mutex = PTHREAD_MUTEX_INITIALIZER;
