+++ src/win32/strptime.c
@@ -44,11 +44,7 @@ __RCSID("$NetBSD: strptime.c,v 1.36 2012/03/13 21:13:48 christos Exp $");
 #include <stdint.h>
 #include "../framework/mlt_api.h"
 #ifdef _MSC_VER
-	#include <string.h>
-
-	#define strncasecmp _strnicmp
-	#define strcasecmp  _stricmp 
-	#define tzname      _tzname
+#include "../framework/msvc_posix_compat.h"
 #endif
 /*
 #include <tzfile.h>
